requirements.txt
black==26.1.0
bleach==6.3.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
idna==3.11
kaggle==1.8.3
kagglesdk==0.1.15
librt==0.7.8
mypy==1.19.1
mypy_extensions==1.1.0
numpy==2.4.2
packaging==26.0
pandas==3.0.0
pathspec==1.0.4
platformdirs==4.5.1
protobuf==6.33.5
pyarrow==23.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytokens==0.4.1
requests==2.32.5
six==1.17.0
text-unidecode==1.3
tqdm==4.67.3
types-requests==2.32.4.20260107
types-tqdm==4.67.2.20260202
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
webencodings==0.5.1
